requirements.txt
numpy==1.17.4
pandas==0.25.3
scikit-learn==0.22
matplotlib==3.1.2
regressors==0.0.3
Flask==1.1.1
python-dotenv==0.10.3
